refactoring home page and error page and end state page to be SimpleNavigationPage.
