+++ cmd/samples/pageflow/src/components/LoadingPage.js
