+++ cmd/samples/pageflow/src/components/index.js
@@ -1,4 +1,6 @@
 export { default as Button } from './Button';
-export { default as EndStatePage } from './EndStatePage';
+export { default as ErrorPage } from './ErrorPage';
 export { default as LinkButton } from './LinkButton';
+export { default as LoadingPage } from './LoadingPage';
 export { default as ProgressButton } from './ProgressButton';
+export { default as SimpleNavigationPage } from './SimpleNavigationPage';
