+++ cmd/samples/pageflow/src/routes/index.js
@@ -1,7 +1,4 @@
 export { default as CreateProductPage } from './CreateProductPage';
-export { default as ErrorPage } from './ErrorPage';
-export { default as HomePage } from './HomePage';
-export { default as LoadingPage } from './LoadingPage';
 export { default as ProductPage } from './ProductPage';
 export { default as ProductReviewPage } from './ProductReviewPage';
 export { default as ProductSplitViewPage } from './ProductSplitViewPage';
